+++ src/utils/constants.ts
@@ -7,3 +7,5 @@ export const DEFAULT_REDIS_PORT = 6379
 // pagination related constants
 export const DEFAULT_OFFSET = 1
 export const DEFAULT_LIMIT = 10
+
+export const DEFAULT_CRON_TIME = '*/10 * * * * *'
